backend/pyproject.toml
[tool.poetry]
name = "wschat-backend"
version = "0.1.0"
description = ""
authors = ["B4lex <b4lex@github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.7"
websockets = "10.0"
channels = "3.0.4"
channels-redis = "3.3.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
